+++ app/src/main/java/com/github/orelzion/spotifyclone/view/AlbumListFragment.kt
@@ -7,12 +7,14 @@ import android.widget.ProgressBar
 import android.widget.Toast
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.commit
 import androidx.recyclerview.widget.RecyclerView
 import com.github.orelzion.spotifyclone.R
 import com.github.orelzion.spotifyclone.model.AlbumsResponse
 import com.github.orelzion.spotifyclone.model.repository.BrowseRepository
 import com.github.orelzion.spotifyclone.viewmodel.AlbumViewData
+import com.github.orelzion.spotifyclone.viewmodel.AlbumsViewModel
 
 
 /**
@@ -22,6 +24,8 @@ class AlbumListFragment : Fragment(R.layout.fragment_items_list) {
 
     private val albumsListView: RecyclerView by lazy { requireView().findViewById(R.id.itemsList) }
     private val albumsAdapter: AlbumsAdapter by lazy { AlbumsAdapter() }
+
+    private val albumsViewModel by activityViewModels<AlbumsViewModel>()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -61,25 +65,11 @@ class AlbumListFragment : Fragment(R.layout.fragment_items_list) {
     }
 
     private fun loadAlbums() {
-        BrowseRepository.fetchNewReleases { response, t ->
-            if (response != null) {
-                albumsAdapter.submitList(responseToAlbumsData(response))
-            } else {
-                Toast.makeText(requireContext(), R.string.loading_error, Toast.LENGTH_LONG).show()
-                Log.e(AlbumListFragment::javaClass.name, "Failed to load albums data", t)
-            }
-        }
+        albumsViewModel.loadAlbums()
+
+        albumsViewModel.bindViewDAta().observe(viewLifecycleOwner, {
+            albumsAdapter.submitList(it)
+        })
     }
 
-    private fun responseToAlbumsData(response: AlbumsResponse): List<AlbumViewData> {
-        return response.items.map {
-            AlbumViewData(
-                id = it.id,
-                imageUrl = it.images.first().url,
-                title = it.name,
-                artistName = it.artists.first().name,
-                totalTracks = it.totalTracks
-            )
-        }
-    }
 }
